requirements.txt
Pillow
keras
tensorflow
matplotlib
gitpython
numpy
requests
scipy
# Plotting --------------------------------------------------------------------

